requirements.txt
flask
pandas
spacy
pypdf
python-docx
scikit-learn
openpyxl
werkzeug
